lib/pyproject.toml
[tool.poetry]
readme = "README.md"
name = "cumo"
version = "0.19.1"
description = "Webブラウザ上に点群を描画する python ライブラリ"
authors = ["Kurusugawa Computer"]
license = "BSD"
keywords = ["point-cloud", "pcd"]
repository = "https://github.com/kurusugawa-computer/cumo"
classifiers = [
	"Development Status :: 3 - Alpha",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: BSD License",
	"Programming Language :: Python",
	"Programming Language :: Python :: 3 :: Only",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Topic :: Utilities",
	"Operating System :: OS Independent",
]
include = [
    "cumo/public/*",
    "cumo/_internal/protobuf/*",
]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
protobuf = "^3.15.6"
websockets = "^8.1"
numpy = "^1.20.2"
Pillow = "^9.0.1"
python-lzf = "^0.2.4"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"
Sphinx = "^4.0.0"
flake8 = "^4.0.1"
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
